+++ Modules/motors/timers.c
@@ -1,7 +1,7 @@
 /*
  ******************************************************************************
  * @file           : timers.c
- * @author         : D. Mucha, K. Czechowicz, A. Rybojad
+ * @author         : D. Mucha
  * @brief          : Timers configuration
  ******************************************************************************
  */
@@ -19,20 +19,6 @@ TIM_HandleTypeDef htim3; // encoder 3 - TIM3
 TIM_HandleTypeDef htim5; // PWM 1,2,3 - TIM5
 TIM_HandleTypeDef htim7; // measuring speed - TIM14
 
-/* number of pulse that encoder count */
-volatile uint16_t enc1PulseNumber;
-volatile uint16_t enc2PulseNumber;
-volatile uint16_t enc3PulseNumber;
-
-/* number of pulse that encoder generates per second */
-volatile int16_t enc1PulsePerSec;
-volatile int16_t enc2PulsePerSec;
-volatile int16_t enc3PulsePerSec;
-
-/* angular value of motors */
-volatile int16_t motor1Velocity;
-volatile int16_t motor2Velocity;
-volatile int16_t motor3Velocity;
 
 void InitTimers() {
 	TIM1_Init();
@@ -182,11 +168,6 @@ void TIM7_Init(void) {
 	HAL_TIM_Base_Start_IT(&htim7);
 }
 
-void TIM7_IRQHandler(void) {
-	HAL_TIM_IRQHandler(&htim7);
-
-}
-
 /* overwritten weak msp functions (gpio and clock config)*/
 
 void HAL_TIM_Encoder_MspInit(TIM_HandleTypeDef *htim_encoder) {
@@ -287,68 +268,4 @@ void HAL_TIM_Base_MspInit(TIM_HandleTypeDef *htim_base) {
 
 }
 
-// @brief interrupt for angular speed calculation
-// @param htim handler for TIM7 interrupt clock
-void HAL_TIM_PeriodElapsedCallback(TIM_HandleTypeDef *htim) {
-	if (htim->Instance == TIM7) {
-
-		// getting actual position of encoder
-		enc1PulseNumber = __HAL_TIM_GET_COUNTER(&htim1);
-		enc2PulseNumber = __HAL_TIM_GET_COUNTER(&htim2);
-		enc3PulseNumber = __HAL_TIM_GET_COUNTER(&htim3);
-
-		// checking if direction is negative
-		// if position decreases from high value then direction is negative
-		// uint16_t converted to int16_t and set as negative value
-		if (enc1PulseNumber > ENC1_MAX_PULSE_VALUE / 2) {
-			enc1PulseNumber = ENC1_MAX_PULSE_VALUE - enc1PulseNumber;
-			enc1PulseNumber = (int16_t) enc1PulseNumber;
-			enc1PulseNumber = -1 * enc1PulseNumber;
-		} else
-			(int16_t) enc1PulseNumber;
-
-		if (enc2PulseNumber > ENC2_MAX_PULSE_VALUE / 2) {
-			enc2PulseNumber = ENC2_MAX_PULSE_VALUE - enc2PulseNumber;
-			enc2PulseNumber = (int16_t) enc2PulseNumber;
-			enc2PulseNumber = -1 * enc2PulseNumber;
-		} else
-			(int16_t) enc2PulseNumber;
-
-		if (enc3PulseNumber > ENC3_MAX_PULSE_VALUE / 2) {
-			enc3PulseNumber = ENC3_MAX_PULSE_VALUE - enc3PulseNumber;
-			enc3PulseNumber = (int16_t) enc3PulseNumber;
-			enc3PulseNumber = -1 * enc3PulseNumber;
-		} else
-			(int16_t) enc3PulseNumber;
-
-		// evaluating number of pulses occured in one sec
-		enc1PulsePerSec = enc1PulseNumber * 1000 / VELOCITY_CLOCK_TIME;
-		enc2PulsePerSec = enc2PulseNumber * 1000 / VELOCITY_CLOCK_TIME;
-		enc3PulsePerSec = enc3PulseNumber * 1000 / VELOCITY_CLOCK_TIME;
-
-		// evaluating angular speed basing on encoder resolution
-		motor1Velocity = enc1PulsePerSec / ENC1_PULSE_PER_ROTATION;
-		motor2Velocity = enc2PulsePerSec / ENC2_PULSE_PER_ROTATION;
-		motor3Velocity = enc3PulsePerSec / ENC3_PULSE_PER_ROTATION;
-
-		// resetting counter for next interrupt
-		__HAL_TIM_SET_COUNTER(&htim1, 0);
-		__HAL_TIM_SET_COUNTER(&htim2, 0);
-		__HAL_TIM_SET_COUNTER(&htim3, 0);
-	}
-}
-
-HAL_StatusTypeDef PWM_SetDutyCycle(ChannelType channel, uint16_t duty) {
-	__HAL_TIM_SET_COMPARE(&htim5, channel, duty);
-	return HAL_OK;
-}
-
-void motor_calibration(ChannelType channel) {
-	PWM_SetDutyCycle(channel, 1000);
-	HAL_Delay(5000);
-	PWM_SetDutyCycle(channel, 500);
-	HAL_Delay(5000);
-	PWM_SetDutyCycle(channel, 750);
-	HAL_Delay(5000);
-}
